+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Add Certes.Acme.Resource.Directory to replace Certes.Acme.AcmeDirectory
+- Add support for deleting registration
 
 ## 1.0.7 - 2017-07-20
 ### Changed